+++ test/demo_app/components/dummy/dummy.controller.dart
@@ -1,5 +1,6 @@
 import 'package:momentum/momentum.dart';
 
+import '../basic-list-example/index.dart';
 import 'index.dart';
 
 class DummyController extends MomentumController<DummyModel> {
@@ -9,5 +10,12 @@ class DummyController extends MomentumController<DummyModel> {
       this,
       value: 0,
     );
+  }
+
+  int dummyValue = 0;
+
+  @override
+  void onReady() {
+    dummyValue = controller<BasicListExampleController>().value;
   }
 }